requirements/tune.txt
about-time==4.2.1
alive-progress==3.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.1
attrs==25.4.0
autograd==1.8.0
ax-platform==1.2.4
botorch==0.17.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cma==4.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
deprecated==1.3.1
executing==2.2.1
filelock==3.25.0
fonttools==4.61.1
fsspec==2026.2.0
gpytorch==1.15.2
graphemeu==0.7.2
graphviz==0.21
hydra-core==1.3.2
idna==3.11
imageio==2.37.2
imgaug==0.4.0
ipython==9.11.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyterlab-widgets==3.0.16
kiwisolver==1.4.9
lazy-loader==0.5
linear-operator==0.6.1
lmdb==1.7.5
markdown==3.10.2
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
moocore==0.2.0
mpmath==1.3.0
msgpack==1.1.2
multipledispatch==1.0.0
mypy-extensions==1.1.0
narwhals==2.17.0
networkx==3.6.1
numpy==2.4.2
omegaconf==2.3.0
opencv-python==4.13.0.92
opt-einsum==3.4.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
prompt-toolkit==3.0.52
protobuf==7.34.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pymoo==0.6.1.6
pyparsing==3.3.2
pyre-extensions==0.0.32
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
pyyaml==6.0.3
ray==2.54.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
shapely==2.1.2
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tensorboardx==2.6.4
threadpoolctl==3.6.0
tifffile==2026.3.3
tqdm==4.67.3
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.3
wcwidth==0.6.0
widgetsnbextension==4.0.15
wrapt==2.1.2
# The following packages are considered to be unsafe in a requirements file:
setuptools==82.0.0
